Separate out `useGetRolesList` hook
