+++ features/admin.console-settings.v1/hooks/use-console-roles.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2023-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import { useRolesList } from "@wso2is/admin.roles.v2/api/roles";
+import useGetRolesList from "@wso2is/admin.roles.v2/api/use-get-roles-list";
 import useConsoleSettings from "./use-console-settings";
 
 /**
@@ -45,7 +45,7 @@ const useConsoleRoles = (
         isLoading: isConsoleRolesFetchRequestLoading,
         error: consoleRolesFetchRequestError,
         mutate: mutateConsoleRolesFetchRequest
-    } = useRolesList(
+    } = useGetRolesList(
         count,
         startIndex,
         `audience.value eq ${consoleId}${ filter ? ` and ${ filter }` : "" }`,
