+++ features/admin.roles.v2/components/edit-role/edit-role-basic.tsx
@@ -25,7 +25,8 @@ import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
 import { Dispatch } from "redux";
 import { Divider } from "semantic-ui-react";
-import { deleteRoleById, updateRoleDetails, useRolesList } from "../../api";
+import { deleteRoleById, updateRoleDetails } from "../../api";
+import useGetRolesList from "../../api/use-get-roles-list";
 import { RoleAudienceTypes, RoleConstants, Schemas } from "../../constants";
 import { PatchRoleDataInterface, RoleBasicInterface, RoleEditSectionsInterface } from "../../models/roles";
 import { RoleDeleteErrorConfirmation } from "../wizard/role-delete-error-confirmation";
@@ -72,7 +73,7 @@ export const BasicRoleDetails: FunctionComponent<BasicRoleProps> = (props: Basic
         isLoading: isRolesListLoading,
         error: rolesListError,
         isValidating: isRolesListValidating
-    } = useRolesList(undefined, undefined, roleNameSearchQuery, "users,groups,permissions,associatedApplications");
+    } = useGetRolesList(undefined, undefined, roleNameSearchQuery, "users,groups,permissions,associatedApplications");
 
     /**
      * Dispatches the alert object to the redux store.
