+++ features/admin.validation.v1/pages/validation-config-edit.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2023-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -21,7 +21,7 @@ import { useRequiredScopes } from "@wso2is/access-control";
 import { AppConstants, AppState, FeatureConfigInterface, history } from "@wso2is/admin.core.v1";
 import { serverConfigurationConfig } from "@wso2is/admin.extensions.v1";
 import { useGroupList } from "@wso2is/admin.groups.v1/api";
-import { useRolesList } from "@wso2is/admin.roles.v2/api";
+import useGetRolesList from "@wso2is/admin.roles.v2/api/use-get-roles-list";
 import {
     ConnectorPropertyInterface,
     GovernanceConnectorInterface,
@@ -171,7 +171,7 @@ export const ValidationConfigEditPage: FunctionComponent<MyAccountSettingsEditPa
         data: currentRoleList,
         isLoading: isRolesListLoading,
         error: rolesListError
-    } = useRolesList(
+    } = useGetRolesList(
         rolesListItemLimit,
         roleListOffset,
         null,
